+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Updates the `fontSize` and `padding` attributes in the single variant of the `Accordion` component.
 - Updates all QA urls from Tugboat QA to Vercel.
+- Updates the `Logo` component to include new variants for `Apple App Store`, `Clever Badge` and `Google Play`.
 
 ### Fixes